Added first draft of collision physics and enabled warnings as errors

+++ Ox93/Ox93/Game/Menu/Ox93_MenuButton.cpp
@@ -139,5 +139,5 @@ void Ox93_MenuButton::SetPosition(int iX, int iY)
 	{
 		m_pxImage->SetPosition(iX, iY);
 	}
-	m_xPosition = Ox93_Vector_2(iX, iY);
+	m_xPosition = Ox93_Vector_2((float)iX, (float)iY);
 }
